python_detector/requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
Pillow
